requirements.txt
mlx
transformers
huggingface-hub
Pillow
numpy
opencv-python
torch
